fix: use uint32_t/int64_t to match Rust FFI signatures (#68)

+++ poc/baremetal_riscv/simulation/stubs.c
@@ -5,14 +5,16 @@
  * in Rust FFI. These stubs allow the C test to link and run under Spike.
  */
 
+#include <stdint.h>
+
 /* Constraint function pointer type matching Rust's ConstraintFn */
-typedef int (*constraint_fn)(long long *);
+typedef uint32_t (*constraint_fn)(int64_t *);
 
 /*
  * compose_and_fast: logical AND of two constraints.
  * Referenced via la in observe_full.S run_narrow.
  */
-int compose_and_fast(constraint_fn fa, constraint_fn fb, long long *coord)
+uint32_t compose_and_fast(constraint_fn fa, constraint_fn fb, int64_t *coord)
 {
     return fa(coord) && fb(coord);
 }
@@ -21,7 +23,7 @@ int compose_and_fast(constraint_fn fa, constraint_fn fb, long long *coord)
  * compose_or_fast: logical OR of two constraints.
  * Referenced via la in observe_full.S run_broad.
  */
-int compose_or_fast(constraint_fn fa, constraint_fn fb, long long *coord)
+uint32_t compose_or_fast(constraint_fn fa, constraint_fn fb, int64_t *coord)
 {
     return fa(coord) || fb(coord);
 }
